Liste eingefügt

+++ PR2.Pflichuebung1-Polynome/src/hsMannheim/pr2/ws2018/grp7/pflichtuebung1/LinearList.java
@@ -6,7 +6,7 @@ public class LinearList {
 	
 	public ListElement first;
 	
-	public void addFirst(int value) {
+	public void addFirst(double value) {
 	
 		ListElement n = new ListElement();
